requirements.txt
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1

